(PE-30014) Fix up installer_team_release_creation_job
This modifies the installer_team_release_creation_job script to work with the recent changes in pe-integration.yaml and pe-modules.yaml, and adds pe-backup-tools and pe_installer. It also changes CODENAME to MAINLINE_BRANCH to add some clarity to what that variable is mostly being used for, changes PE_FAMILY to X_Y for additional clarity, adds some extra comments to the yaml files, sets kickoff times for pipelines and disables mainline CI pipelines (except for main), and all changes are now in a single commit instead of several.