(maint) Force push temp branch in move_ci_pipeline_kickoff
If a previous PR failed to merge automatically, it would leave the temp branch in place on ci-job-configs, and a subsequent run of this script would fail.  This force pushes the temp branch so even if it already exists, it will be modified for the latest run of the script.