(maint) Fix move_ci_pipline_kickoff merge error
Because ci-job-configs requires a review on all PRs, when hub tries to hit the merge endpoint, it will think it failed.  However, puppetlabs-jenkins has admin on the repo, and it appears the merge actually does happen. So this allows for the failure exit code, and just double checks that the merge actually happened afterwards.