(maint) More fixes for move_ci_pipeline_kickoff
There seems to be some weird thing around deleting the branch. Sometimes it can't find the branch on the remote. So this adds a sleep and does a fetch to ensure we are in sync. If the branch still doesn't exist, then maybe the branch got auto-deleted on merge, so we don't try pushing a branch deletion. Also, failing to delete the branch isn't fatal, since we force push every time this script is run, so just exit 0 if we can't delete it for whatever reason.